+++ stdlib/InteractiveUtils/src/InteractiveUtils.jl
@@ -45,7 +45,7 @@ end
 varinfo(pat::Regex) = varinfo(Main, pat)
 
 """
-    versioninfo(io::IO=STDOUT; verbose::Bool=false, packages::Bool=false)
+    versioninfo(io::IO=stdout; verbose::Bool=false, packages::Bool=false)
 
 Print information about the version of Julia in use. The output is
 controlled with boolean keyword arguments:
@@ -53,7 +53,7 @@ controlled with boolean keyword arguments:
 - `packages`: print information about installed packages
 - `verbose`: print all additional information
 """
-function versioninfo(io::IO=STDOUT; verbose::Bool=false, packages::Bool=false)
+function versioninfo(io::IO=stdout; verbose::Bool=false, packages::Bool=false)
     println(io, "Julia Version $VERSION")
     if !isempty(Base.GIT_VERSION_INFO.commit_short)
         println(io, "Commit $(Base.GIT_VERSION_INFO.commit_short) ($(Base.GIT_VERSION_INFO.date_string))")
@@ -68,7 +68,7 @@ function versioninfo(io::IO=STDOUT; verbose::Bool=false, packages::Bool=false)
     if verbose
         lsb = ""
         if Sys.islinux()
-            try lsb = readchomp(pipeline(`lsb_release -ds`, stderr=DevNull)) end
+            try lsb = readchomp(pipeline(`lsb_release -ds`, stderr=devnull)) end
         end
         if Sys.iswindows()
             try lsb = strip(read(`$(ENV["COMSPEC"]) /c ver`, String)) end