+++ stdlib/REPL/test/TerminalMenus/runtests.jl
@@ -6,7 +6,7 @@ TerminalMenus.config(supress_output=true)
 
 function simulateInput(expectedResult, menu::TerminalMenus.AbstractMenu, keys...)
     # If we cannot write to the buffer, skip the test
-    !(:buffer in fieldnames(typeof(STDIN))) && return true
+    !(:buffer in fieldnames(typeof(stdin))) && return true
 
     keydict =  Dict(:up => "\e[A",
                     :down => "\e[B",
@@ -14,9 +14,9 @@ function simulateInput(expectedResult, menu::TerminalMenus.AbstractMenu, keys...
 
     for key in keys
         if isa(key, Symbol)
-            write(STDIN.buffer, keydict[key])
+            write(stdin.buffer, keydict[key])
         else
-            write(STDIN.buffer, "$key")
+            write(stdin.buffer, "$key")
         end
     end