+++ test/cmdlineargs.jl
@@ -18,7 +18,7 @@ end
 function readchomperrors(exename::Cmd)
     out = Base.PipeEndpoint()
     err = Base.PipeEndpoint()
-    p = spawn(exename, DevNull, out, err)
+    p = spawn(exename, devnull, out, err)
     o = @async(readchomp(out))
     e = @async(readchomp(err))
     return (success(p), fetch(o), fetch(e))
@@ -173,19 +173,19 @@ let exename = `$(Base.julia_cmd()) --sysimage-native-code=yes --startup-file=no`
 
     # -g
     @test readchomp(`$exename -E "Base.JLOptions().debug_level" -g`) == "2"
-    let code = read(`$exename -g0 -i -e "code_llvm(STDOUT, +, (Int64, Int64), false, true); exit()"`, String)
+    let code = read(`$exename -g0 -i -e "code_llvm(stdout, +, (Int64, Int64), false, true); exit()"`, String)
         @test contains(code, "llvm.module.flags")
         @test !contains(code, "llvm.dbg.cu")
         @test !contains(code, "int.jl")
         @test !contains(code, "Int64")
     end
-    let code = read(`$exename -g1 -i -e "code_llvm(STDOUT, +, (Int64, Int64), false, true); exit()"`, String)
+    let code = read(`$exename -g1 -i -e "code_llvm(stdout, +, (Int64, Int64), false, true); exit()"`, String)
         @test contains(code, "llvm.module.flags")
         @test contains(code, "llvm.dbg.cu")
         @test contains(code, "int.jl")
         @test !contains(code, "Int64")
     end
-    let code = read(`$exename -g2 -i -e "code_llvm(STDOUT, +, (Int64, Int64), false, true); exit()"`, String)
+    let code = read(`$exename -g2 -i -e "code_llvm(stdout, +, (Int64, Int64), false, true); exit()"`, String)
         @test contains(code, "llvm.module.flags")
         @test contains(code, "llvm.dbg.cu")
         @test contains(code, "int.jl")
@@ -275,7 +275,7 @@ let exename = `$(Base.julia_cmd()) --sysimage-native-code=yes --startup-file=no`
     # test passing arguments
     mktempdir() do dir
         testfile = joinpath(dir, tempname())
-        # write a julia source file that just prints ARGS to STDOUT
+        # write a julia source file that just prints ARGS to stdout
         write(testfile, """
             println(ARGS)
             """)
@@ -387,10 +387,10 @@ let exename = joinpath(Sys.BINDIR, Base.julia_exename()),
             joinpath(@__DIR__, "nonexistent"),
             "$sysname.nonexistent",
             )
-        let stderr = Pipe(),
-            p = spawn(pipeline(`$exename --sysimage=$nonexist_image`, stderr=stderr))
-            close(stderr.in)
-            let s = read(stderr, String)
+        let err = Pipe(),
+            p = spawn(pipeline(`$exename --sysimage=$nonexist_image`, stderr=err))
+            close(err.in)
+            let s = read(err, String)
                 @test contains(s, "ERROR: could not load library \"$nonexist_image\"\n")
                 @test !contains(s, "Segmentation fault")
                 @test !contains(s, "EXCEPTION_ACCESS_VIOLATION")
@@ -400,10 +400,10 @@ let exename = joinpath(Sys.BINDIR, Base.julia_exename()),
             @test p.exitcode == 1
         end
     end
-    let stderr = Pipe(),
-        p = spawn(pipeline(`$exename --sysimage=$libjulia`, stderr=stderr))
-        close(stderr.in)
-        let s = read(stderr, String)
+    let err = Pipe(),
+        p = spawn(pipeline(`$exename --sysimage=$libjulia`, stderr=err))
+        close(err.in)
+        let s = read(err, String)
             @test s == "ERROR: System image file failed consistency check: maybe opened the wrong version?\n"
         end
         @test !success(p)
@@ -429,7 +429,7 @@ let exename = `$(Base.julia_cmd()) --sysimage-native-code=yes`
 end
 
 # Make sure `julia --lisp` doesn't break
-run(pipeline(DevNull, `$(joinpath(Sys.BINDIR, Base.julia_exename())) --lisp`, DevNull))
+run(pipeline(devnull, `$(joinpath(Sys.BINDIR, Base.julia_exename())) --lisp`, devnull))
 
 # Test that `julia [some other option] --lisp` is disallowed
 @test readchomperrors(`$(joinpath(Sys.BINDIR, Base.julia_exename())) -Cnative --lisp`) ==